+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.4.0
 
+* Adding the `stemmers/french/porter` namespace.
+* Adding the `phonetics/french/soundex` namespace.
+* Adding the `phonetics/french/soundex2` namespace.
+* Adding the `phonetics/french/phonex` namespace.
 * Moving the Cologne phonetic algorithm to the `phonetics/german/cologne` namespace.
 
 ## 0.3.0