+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.5.0
+
+* Adding the `phonetics/german/phonem` namespace.
+
 ## 0.4.0
 
 * Adding the `phonetics/french/phonetic` namespace.